Separación de interfaz de telegram para usuaries
